@@ -14,26 +14,55 @@ import (
 	"github.com/maxlaverse/terraform-provider-bitwarden/internal/bitwarden/crypto/symmetrickey"
 )
 
-func Encrypt(plainValue []byte, key symmetrickey.Key) (string, error) {
+var (
+	SafeMode = true
+)
+
+func EncryptAsString(plainValue []byte, key symmetrickey.Key) (string, error) {
+	res, err := Encrypt(plainValue, key)
+	if err != nil {
+		return "", err
+	}
+	return res.String(), nil
+}
+
+func Encrypt(plainValue []byte, key symmetrickey.Key) (*encryptedstring.EncryptedString, error) {
+	if len(plainValue) == 0 {
+		return nil, fmt.Errorf("trying to encrypt nothing")
+	}
 	randomIV := make([]byte, 16)
-	rand.Read(randomIV)
+	_, err := rand.Read(randomIV)
+	if err != nil {
+		return nil, fmt.Errorf("error generating random bytes: %w", err)
+	}
 
 	if len(key.EncryptionKey) == 0 {
-		return "", fmt.Errorf("no encryption key was provided: %v", key.EncryptionKey)
+		return nil, fmt.Errorf("no encryption key was provided: %v", key.EncryptionKey)
 	}
 	if len(randomIV) != 16 {
-		return "", fmt.Errorf("bad IV length - expected 16, got: %d", len(randomIV))
+		return nil, fmt.Errorf("bad IV length - expected 16, got: %d", len(randomIV))
 	}
 
 	data, err := aes256Encode(plainValue, key.EncryptionKey, randomIV, 16)
 	if err != nil {
-		return "", fmt.Errorf("error to aes256encoding data: %w", err)
+		return nil, fmt.Errorf("error to aes256encoding data: %w", err)
 	}
 
 	hmac := hmacSum(append(randomIV, data...), key.MacKey, sha256.New)
 
 	res := encryptedstring.New(randomIV, data, hmac, key)
-	return res.String(), nil
+
+	if SafeMode {
+		safeDecryptedValue, err := Decrypt(&res, &key)
+		if err != nil {
+			return nil, fmt.Errorf("error reversing decryption (safe mode): %w", err)
+		}
+		if !bytes.Equal(safeDecryptedValue, plainValue) {
+			return nil, fmt.Errorf("failed to reverse decryption (safe mode)")
+		}
+	}
+
+	return &res, nil
 }
 
 func DecryptPrivateKey(encryptedPrivateKeyStr string, encryptionKey symmetrickey.Key) (*rsa.PrivateKey, error) {
@@ -42,7 +71,7 @@ func DecryptPrivateKey(encryptedPrivateKeyStr string, encryptionKey symmetrickey
 		return nil, fmt.Errorf("error decrypting private key: %w", err)
 	}
 
-	decryptedPrivateKey, err := decrypt(encString, &encryptionKey)
+	decryptedPrivateKey, err := Decrypt(encString, &encryptionKey)
 	if err != nil {
 		return nil, fmt.Errorf("error decrypting private key: %w", err)
 	}
@@ -61,17 +90,14 @@ func DecryptEncryptionKey(encryptedKeyStr string, key symmetrickey.Key) (*symmet
 		return nil, fmt.Errorf("error decrypting encryption key: %w", err)
 	}
 	if encKeyCipher.Key.EncryptionType == symmetrickey.AesCbc256_B64 {
-		decEncKey, err = decrypt(encKeyCipher, &key)
+		decEncKey, err = Decrypt(encKeyCipher, &key)
 		if err != nil {
 			return nil, fmt.Errorf("error decrypting encryption key: %w", err)
 		}
 	} else if encKeyCipher.Key.EncryptionType == symmetrickey.AesCbc256_HmacSha256_B64 {
-		newKey, err := key.StretchKey()
-		if err != nil {
-			return nil, fmt.Errorf("error stretching key encryption key: %w", err)
-		}
+		newKey := key.StretchKey()
 
-		decEncKey, err = decrypt(encKeyCipher, newKey)
+		decEncKey, err = Decrypt(encKeyCipher, &newKey)
 		if err != nil {
 			return nil, fmt.Errorf("error decrypting encryption key: %w", err)
 		}
@@ -86,7 +112,7 @@ func DecryptEncryptionKey(encryptedKeyStr string, key symmetrickey.Key) (*symmet
 	return encryptionKey, nil
 }
 
-func decrypt(encString *encryptedstring.EncryptedString, key *symmetrickey.Key) ([]byte, error) {
+func Decrypt(encString *encryptedstring.EncryptedString, key *symmetrickey.Key) ([]byte, error) {
 	if encString.Key.EncryptionType == symmetrickey.AesCbc128_HmacSha256_B64 && key.EncryptionType == symmetrickey.AesCbc256_B64 {
 		return nil, fmt.Errorf("unsupported old scheme")
 	}
@@ -99,9 +125,13 @@ func decrypt(encString *encryptedstring.EncryptedString, key *symmetrickey.Key) 
 		return nil, fmt.Errorf("hmac value is missing")
 	}
 
-	hmac := hmacSum(append(encString.IV, encString.Data...), key.MacKey, sha256.New)
-	if !bytes.Equal(hmac, encString.Hmac) {
-		return nil, fmt.Errorf("hmac comparison failed: %s!=%s", encString.Hmac, hmac)
+	if len(encString.Hmac) != len(key.MacKey) {
+		return nil, fmt.Errorf("hmac lengths differ: %d!=%d", len(encString.Hmac), len(key.MacKey))
+	}
+
+	computedHmac := hmacSum(append(append([]byte{}, encString.IV...), encString.Data...), key.MacKey, sha256.New)
+	if !bytes.Equal(computedHmac, encString.Hmac) {
+		return nil, fmt.Errorf("hmac comparison failed: %v != %v", computedHmac, encString.Hmac)
 	}
 	decData, err := aes256Decode(encString.Data, key.EncryptionKey, encString.IV)
 	if err != nil {
