@@ -2,13 +2,14 @@ package models
 
 import (
 	"errors"
-	"strings"
 	"time"
 )
 
 var (
-	ErrObjectNotFound     = errors.New("object not found")
-	ErrAttachmentNotFound = errors.New("attachment not found")
+	ErrObjectNotFound      = errors.New("object not found")
+	ErrAttachmentNotFound  = errors.New("attachment not found")
+	ErrVaultLocked         = errors.New("vault is locked")
+	ErrWrongMasterPassword = errors.New("invalid master password")
 )
 
 type ItemType int
@@ -16,10 +17,6 @@ type ItemType int
 const (
 	ItemTypeLogin      ItemType = 1
 	ItemTypeSecureNote ItemType = 2
-)
-
-const (
-	DefaultBitwardenServerURL = "https://vault.bitwarden.com"
 )
 
 type FieldType int
@@ -34,42 +31,24 @@ const (
 type ObjectType string
 
 const (
-	ObjectTypeItem          ObjectType = "item"
-	ObjectTypeAttachment    ObjectType = "attachment"
-	ObjectTypeFolder        ObjectType = "folder"
-	ObjectTypeOrgCollection ObjectType = "org-collection"
-	ObjectTypeOrganization  ObjectType = "organization"
+	ObjectTypeItem                ObjectType = "item"
+	ObjectTypeAttachment          ObjectType = "attachment"
+	ObjectTypeFolder              ObjectType = "folder"
+	ObjectTypeOrgCollection       ObjectType = "org-collection"
+	ObjectTypeOrganization        ObjectType = "organization"
+	ObjectTypeList                ObjectType = "list"              // encapsulates collection list response
+	ObjectTypeCollectionDetails   ObjectType = "collectionDetails" // collection listed in sync
+	ObjectTypeCollection          ObjectType = "collection"        // used when refetching collections
+	ObjectTypeProfile             ObjectType = "profile"
+	ObjectTypeSync                ObjectType = "sync"
+	ObjectTypeProfileOrganization ObjectType = "profileOrganization"   // organization under profile
+	ObjectCipherDetails           ObjectType = "cipherDetails"         // when creating attachment data
+	ObjectAttachmentFileUpload    ObjectType = "attachment-fileUpload" // when creating attachment data
 )
-
-type VaultStatus string
 
 const (
-	StatusLocked          VaultStatus = "locked"
-	StatusUnauthenticated VaultStatus = "unauthenticated"
-	StatusUnlocked        VaultStatus = "unlocked"
+	DateLayout = "2006-01-02T15:04:05.000Z"
 )
-
-type Status struct {
-	ServerURL string      `json:"serverURL,omitempty"`
-	LastSync  time.Time   `json:"lastSync,omitempty"`
-	UserEmail string      `json:"userEmail,omitempty"`
-	UserID    string      `json:"userID,omitempty"`
-	Status    VaultStatus `json:"status,omitempty"`
-}
-
-func (s *Status) VaultFromServer(serverUrl string) bool {
-	providerServerUrl := trimSlashSuffix(serverUrl)
-	vaultServerUrl := trimSlashSuffix(s.ServerURL)
-	return vaultServerUrl == providerServerUrl || len(vaultServerUrl) == 0 && providerServerUrl == DefaultBitwardenServerURL
-}
-
-func (s *Status) VaultOfUser(email string) bool {
-	return s.UserEmail == email
-}
-
-func trimSlashSuffix(serverUrl string) string {
-	return strings.TrimSuffix(serverUrl, "/")
-}
 
 type Login struct {
 	Username string     `json:"username,omitempty"`
@@ -89,6 +68,10 @@ const (
 	URIMatchNever      URIMatch = 5
 )
 
+func (u URIMatch) ToPointer() *URIMatch {
+	return &u
+}
+
 type LoginURI struct {
 	Match *URIMatch `json:"match,omitempty"`
 	URI   string    `json:"uri,omitempty"`
@@ -99,30 +82,37 @@ type SecureNote struct {
 }
 
 type Object struct {
-	CollectionIds  []string      `json:"collectionIds,omitempty"`
-	CreationDate   *time.Time    `json:"creationDate,omitempty"`
-	DeletedDate    *time.Time    `json:"deletedDate,omitempty"`
-	ID             string        `json:"id,omitempty"`
-	ExternalID     string        `json:"externalId,omitempty"`
-	FolderID       string        `json:"folderId,omitempty"`
-	Groups         []interface{} `json:"groups"` // Not implemented yet
-	Login          Login         `json:"login,omitempty"`
-	Name           string        `json:"name,omitempty"`
-	Notes          string        `json:"notes,omitempty"`
-	Object         ObjectType    `json:"object,omitempty"`
-	OrganizationID string        `json:"organizationId,omitempty"`
-	SecureNote     SecureNote    `json:"secureNote,omitempty"`
-	Type           ItemType      `json:"type,omitempty"`
-	Fields         []Field       `json:"fields,omitempty"`
-	Reprompt       int           `json:"reprompt,omitempty"`
-	Favorite       bool          `json:"favorite,omitempty"`
-	RevisionDate   *time.Time    `json:"revisionDate,omitempty"`
-	Attachments    []Attachment  `json:"attachments,omitempty"`
+	Attachments         []Attachment          `json:"attachments,omitempty"`
+	Card                []byte                `json:"-"`
+	CollectionIds       []string              `json:"collectionIds,omitempty"`
+	CreationDate        *time.Time            `json:"creationDate,omitempty"`
+	DeletedDate         *time.Time            `json:"deletedDate,omitempty"`
+	Edit                bool                  `json:"edit,omitempty"`
+	Favorite            bool                  `json:"favorite,omitempty"`
+	Fields              []Field               `json:"fields,omitempty"`
+	FolderID            string                `json:"folderId,omitempty"`
+	Groups              []interface{}         `json:"groups"` // To be kept for the CLI when creating org-collections
+	ID                  string                `json:"id,omitempty"`
+	Identity            string                `json:"-"`
+	Key                 string                `json:"key,omitempty"`
+	Login               Login                 `json:"login,omitempty"`
+	Name                string                `json:"name,omitempty"`
+	Notes               string                `json:"notes,omitempty"`
+	Object              ObjectType            `json:"object,omitempty"`
+	OrganizationID      string                `json:"organizationId,omitempty"`
+	OrganizationUseTotp bool                  `json:"organizationUseTotp,omitempty"`
+	PasswordHistory     []PasswordHistoryItem `json:"passwordHistory,omitempty"`
+	Reprompt            int                   `json:"reprompt,omitempty"`
+	RevisionDate        *time.Time            `json:"revisionDate,omitempty"`
+	SecureNote          SecureNote            `json:"secureNote,omitempty"`
+	Type                ItemType              `json:"type,omitempty"`
+	ViewPassword        bool                  `json:"viewPassword,omitempty"`
 }
 
-const (
-	DateLayout = "2006-01-02T15:04:05.000Z"
-)
+type PasswordHistoryItem struct {
+	LastUsedDate *time.Time `json:"lastUsedDate,omitempty"`
+	Password     string     `json:"password,omitempty"`
+}
 
 type Field struct {
 	Name     string    `json:"name,omitempty"`
@@ -132,9 +122,11 @@ type Field struct {
 }
 
 type Attachment struct {
-	ID       string `json:"id,omitempty"`
-	FileName string `json:"fileName,omitempty"`
-	Size     string `json:"size,omitempty"`
-	SizeName string `json:"sizeName,omitempty"`
-	Url      string `json:"url,omitempty"`
+	ID       string     `json:"id,omitempty"`
+	FileName string     `json:"fileName,omitempty"`
+	Size     string     `json:"size,omitempty"`
+	SizeName string     `json:"sizeName,omitempty"`
+	Url      string     `json:"url,omitempty"`
+	Key      string     `json:"key"`
+	Object   ObjectType `json:"object"`
 }
