@@ -20,15 +20,22 @@ func TestAccResourceOrgCollection(t *testing.T) {
 		Steps: []resource.TestStep{
 			{
 				ResourceName: resourceName,
-				Config:       tfConfigProvider() + tfConfigResourceOrgCollection(),
+				Config:       tfConfigProvider() + tfConfigResourceOrgCollection("org-col-bar"),
 				Check: resource.ComposeTestCheckFunc(
-					resource.TestMatchResourceAttr(
-						resourceName, attributeName, regexp.MustCompile("^org-col-bar$"),
+					resource.TestCheckResourceAttr(
+						resourceName, attributeName, "org-col-bar",
 					),
 					resource.TestMatchResourceAttr(
 						resourceName, attributeID, regexp.MustCompile(regExpId),
 					),
 					getObjectID(resourceName, &objectID),
+				),
+			},
+			{
+				ResourceName: resourceName,
+				Config:       tfConfigProvider() + tfConfigResourceOrgCollection("org-col-new-name-bar"),
+				Check: resource.TestCheckResourceAttr(
+					resourceName, attributeName, "org-col-new-name-bar",
 				),
 			},
 			{
@@ -52,14 +59,14 @@ func orgCollectionImportID(resourceName string) func(s *terraform.State) (string
 	}
 }
 
-func tfConfigResourceOrgCollection() string {
+func tfConfigResourceOrgCollection(name string) string {
 	return fmt.Sprintf(`
 	resource "bitwarden_org_collection" "foo_org_col" {
 	provider	= bitwarden
 
 	organization_id = "%s"
 
-	name     = "org-col-bar"
+	name     = "%s"
 }
-`, testOrganizationID)
+`, testOrganizationID, name)
 }
