@@ -82,22 +82,27 @@ const (
 	descriptionRevisionDate           = "Last time the item was updated."
 
 	// Provider field attributes
-	attributeClientID         = "client_id"
-	attributeClientSecret     = "client_secret"
-	attributeEmail            = "email"
-	attributeMasterPassword   = "master_password"
-	attributeServer           = "server"
-	attributeSessionKey       = "session_key"
-	attributeVaultPath        = "vault_path"
-	attributeExtraCACertsPath = "extra_ca_certs"
+	attributeClientID                   = "client_id"
+	attributeClientSecret               = "client_secret"
+	attributeEmail                      = "email"
+	attributeMasterPassword             = "master_password"
+	attributeServer                     = "server"
+	attributeSessionKey                 = "session_key"
+	attributeVaultPath                  = "vault_path"
+	attributeExtraCACertsPath           = "extra_ca_certs"
+	attributeExperimental               = "experimental"
+	attributeExperimentalEmbeddedClient = "embedded_client"
+	attributeDeviceIdentifier           = "device_id"
 
 	// Provider field descriptions
-	descriptionClientSecret     = "Client Secret (env: `BW_CLIENTSECRET`). Do not commit this information in Git unless you know what you're doing. Prefer using a Terraform `variable {}` in order to inject this value from the environment."
-	descriptionClientID         = "Client ID (env: `BW_CLIENTID`)"
-	descriptionEmail            = "Login Email of the Vault (env: `BW_EMAIL`)."
-	descriptionMasterPassword   = "Master password of the Vault (env: `BW_PASSWORD`). Do not commit this information in Git unless you know what you're doing. Prefer using a Terraform `variable {}` in order to inject this value from the environment."
-	descriptionServer           = "Bitwarden Server URL (default: `https://vault.bitwarden.com`, env: `BW_URL`)."
-	descriptionSessionKey       = "A Bitwarden Session Key (env: `BW_SESSION`)"
-	descriptionVaultPath        = "Alternative directory for storing the Vault locally (default: `.bitwarden/`, env: `BITWARDENCLI_APPDATA_DIR`)."
-	descriptionExtraCACertsPath = "Extends the well known 'root' CAs (like VeriSign) with the extra certificates in file (env: `NODE_EXTRA_CA_CERTS`)."
+	descriptionClientSecret               = "Client Secret (env: `BW_CLIENTSECRET`). Do not commit this information in Git unless you know what you're doing. Prefer using a Terraform `variable {}` in order to inject this value from the environment."
+	descriptionClientID                   = "Client ID (env: `BW_CLIENTID`)"
+	descriptionEmail                      = "Login Email of the Vault (env: `BW_EMAIL`)."
+	descriptionMasterPassword             = "Master password of the Vault (env: `BW_PASSWORD`). Do not commit this information in Git unless you know what you're doing. Prefer using a Terraform `variable {}` in order to inject this value from the environment."
+	descriptionServer                     = "Bitwarden Server URL (default: `https://vault.bitwarden.com`, env: `BW_URL`)."
+	descriptionSessionKey                 = "A Bitwarden Session Key (env: `BW_SESSION`)"
+	descriptionVaultPath                  = "Alternative directory for storing the Vault locally (default: `.bitwarden/`, env: `BITWARDENCLI_APPDATA_DIR`)."
+	descriptionExtraCACertsPath           = "Extends the well known 'root' CAs (like VeriSign) with the extra certificates in file (env: `NODE_EXTRA_CA_CERTS`)."
+	descriptionExperimental               = "Enable experimental features."
+	descriptionExperimentalEmbeddedClient = "Use the embedded client instead of an external binary."
 )
